fix: Theme context added to switch between light and dark mode
